docker_webapp/app/requirements.txt
appdirs==1.4.3
Django==1.10.6
django-braces==1.11.0
django-cors-middleware==1.3.1
django-debug-toolbar==1.7
django-oauth-toolkit==0.12.0
djangorestframework==3.6.2
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
